requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
py==1.4.27
pytest==2.7.0
wsgiref==0.1.2
sqlite3==3.8.8.3
validator=0.8
